requirements.txt
flask
python-telegram-bot
requests
sqlite3